Refactor generation history into container and composables
